requirements.txt
mss
Pillow
websockets
